requirements.txt
aiogram>=3.13,<4.0
aiosqlite>=0.20,<1.0
opencv-python>=4.10,<5.0
Pillow>=12.0,<13.0
python-dotenv>=1.0,<2.0
pyzbar>=0.1.9,<1.0
qrcode>=7.4,<8.0
reportlab>=4.2,<5.0
SQLAlchemy[asyncio]>=2.0,<3.0
segno>=1.6,<2.0
